+++ crates/rustc_codegen_gpu/tests/codegen/shared.rs
@@ -13,7 +13,7 @@ pub fn alloc_shared(a: &[u8], _a_window: usize, b: &mut [u8], b_window: usize, f
     let mut shared = gpu::GpuShared::<[u8; 10]>::zero();
     let mut chunk_dy_shared = dy_shared.chunk_mut(gpu::MapLinear::new(1));
     let mut chunk_shared = shared.chunk_mut(gpu::MapLinear::new(1));
-    chunk_shared[0] = a[gpu::thread_id::<gpu::DimX>()];
+    chunk_shared[0] = a[gpu::thread_id::<gpu::DimX>() as usize];
     chunk_dy_shared[0] = 1.1;
     gpu::sync_threads();
