+++ crates/rustc_codegen_gpu/tests/codegen_fails/diversed_chunk.rs
@@ -8,7 +8,7 @@
 #[gpu_macros::kernel]
 #[no_mangle]
 pub fn test_diversed_chunk(mut a: &mut [f32]) {
-    let mut local = gpu::GlobalThreadChunk::new(a, gpu::MapLinear::new(gpu::thread_id::<gpu::DimX>())); //~ ERROR Invalid use of diversed data in GPU code
+    let mut local = gpu::GlobalThreadChunk::new(a, gpu::MapLinear::new(gpu::thread_id::<gpu::DimX>() as usize)); //~ ERROR Invalid use of diversed data in GPU code
     //~| ERROR Invalid use of diversed data in GPU code
     local[0] = 1.0;
 }
