+++ crates/rustc_codegen_gpu/tests/codegen_fails/sync_missing_read_write.rs
@@ -9,7 +9,7 @@
 pub fn test_shared_read_write(a: &[u8], b: &mut [u8]) {
     let mut shared = gpu::GpuShared::<[u8; 10]>::zero();
     let mut chunk_shared = shared.chunk_mut(gpu::MapLinear::new(1));
-    chunk_shared[0] = a[gpu::thread_id::<gpu::DimX>()];
+    chunk_shared[0] = a[gpu::thread_id::<gpu::DimX>() as usize];
     let mut chunked_b = gpu::chunk_mut(b, gpu::MapLinear::new(1));
     let val = chunk_shared[0]; // This is safe, as no read/write to other chunk of the shared mem.
     chunked_b[0] = val;
