+++ examples/llm-rs-gpu/tests/gelu.rs
@@ -21,17 +21,16 @@ pub fn gelu_forward_cpu(inp: &[f32]) -> Vec<f32> {
 
 #[test]
 fn test_gelu() {
-    const N: usize = 1024;
-    let h_dinp = random_f32_vec(N);
-    let mut h_doutp = [0.0f32; N];
+    const N: u32 = 1024;
+    let h_dinp = random_f32_vec(N as usize);
+    let mut h_doutp = [0.0f32; N as usize];
     const BLOCK_SIZE: u32 = 256;
     cuda_ctx(0, |ctx, m| {
         let inp = ctx.new_tensor_view(h_dinp.as_slice()).unwrap();
         let mut outp = ctx.new_tensor_view(h_doutp.as_slice()).unwrap();
-        let grid_size = N.div_ceil(BLOCK_SIZE as usize);
-        let config = gpu_host::gpu_config!(grid_size as u32, 1, 1, @const BLOCK_SIZE, 1, 1, 0);
-        gelu_forward_kernel::launch(config, ctx, m, &mut outp, &inp, N as i32)
-            .expect("launch failed");
+        let grid_size = N.div_ceil(BLOCK_SIZE);
+        let config = gpu_host::gpu_config!(grid_size, 1, 1, @const BLOCK_SIZE, 1, 1, 0);
+        gelu_forward_kernel::launch(config, ctx, m, &mut outp, &inp, N).expect("launch failed");
         outp.copy_to_host(&mut h_doutp)
             .expect("copy to host failed");
     });
@@ -63,9 +62,9 @@ pub fn gelu_backward_cpu(inp: &[f32], dout: &[f32]) -> Vec<f32> {
 
 #[test]
 fn test_gelu_backward() {
-    const N: usize = 512;
-    let mut h_inp = random_f32_vec(N);
-    let mut h_dinp = random_f32_vec(N);
+    const N: u32 = 512;
+    let mut h_inp = random_f32_vec(N as usize);
+    let mut h_dinp = random_f32_vec(N as usize);
     h_dinp[0] = -0.000001009;
     h_inp[0] = -0.176_395_06;
     const BLOCK_SIZE: u32 = 128;
@@ -73,9 +72,9 @@ fn test_gelu_backward() {
     cuda_ctx(0, |ctx, m| {
         let inp = ctx.new_tensor_view(h_inp.as_slice()).unwrap();
         let mut dinp = ctx.new_tensor_view(h_dinp.as_slice()).unwrap();
-        let grid_size = N.div_ceil(BLOCK_SIZE as usize);
-        let config = gpu_host::gpu_config!(grid_size as u32, 1, 1, @const BLOCK_SIZE, 1, 1, 0);
-        llm_rs_gpu::gelu_backward_kernel::launch(config, ctx, m, &mut dinp, &inp, N as i32)
+        let grid_size = N.div_ceil(BLOCK_SIZE);
+        let config = gpu_host::gpu_config!(grid_size, 1, 1, @const BLOCK_SIZE, 1, 1, 0);
+        llm_rs_gpu::gelu_backward_kernel::launch(config, ctx, m, &mut dinp, &inp, N)
             .expect("launch failed");
         dinp.copy_to_host(&mut h_dinp).expect("copy to host failed");
     });
